requirements.txt
fastapi
uvicorn
pycryptodome
pyotp
requests